Adding a small check to themeunpacker to make sure the 7z's have the proper structure, fixing image shown even when no archives found.

+++ spruce/scripts/ThemeUnpacker.sh
@@ -8,18 +8,22 @@ ROOT_DIR="/mnt/SDCARD"
 
 cores_online 4
 log_message "Starting theme unpacking process"
-display_text -i "/mnt/SDCARD/spruce/imgs/displayTextPreColor.png" -t "Packed themes detected. Unpacking themes.........." -c dbcda7
 
 # Unpack themes from 7z archives
 for archive in "$THEME_DIR"/*.7z; do
     if [ -f "$archive" ]; then
         theme_name=$(basename "$archive" .7z)
-        7zr x "$archive" -o/
-        if [ $? -eq 0 ]; then
-            rm -f "$archive"
-            log_message "Unpacked and removed: $theme_name.7z"
+        display_text -i "/mnt/SDCARD/spruce/imgs/displayTextPreColor.png" -t "$theme_name packed theme detected. Unpacking.........." -c dbcda7
+        if 7zr l "$archive" | grep -q "/mnt/SDCARD/"; then
+            7zr x "$archive" -o/
+            if [ $? -eq 0 ]; then
+                rm -f "$archive"
+                log_message "Unpacked and removed: $theme_name.7z"
+            else
+                log_message "Failed to unpack: $theme_name.7z"
+            fi
         else
-            log_message "Failed to unpack: $theme_name.7z"
+            log_message "Skipped unpacking: $theme_name.7z (incorrect folder structure)"
         fi
     fi
 done
@@ -30,12 +34,16 @@ RA_FOLDERS_TO_UNPACK="xmb"
 for folder in $RA_FOLDERS_TO_UNPACK; do
     archive="$RA_THEME_DIR/${folder}.7z"
     if [ -f "$archive" ]; then
-         7zr x "$archive" -o/
-        if [ $? -eq 0 ]; then
-            rm -f "$archive"
-            log_message "Unpacked and removed RetroArch folder: ${folder}.7z"
+        if 7zr l "$archive" | grep -q "/mnt/SDCARD/"; then
+            7zr x "$archive" -o/
+            if [ $? -eq 0 ]; then
+                rm -f "$archive"
+                log_message "Unpacked and removed RetroArch folder: ${folder}.7z"
+            else
+                log_message "Failed to unpack RetroArch folder: ${folder}.7z"
+            fi
         else
-            log_message "Failed to unpack RetroArch folder: ${folder}.7z"
+            log_message "Skipped unpacking RetroArch folder: ${folder}.7z (incorrect folder structure)"
         fi
     else
         log_message "RetroArch archive not found: ${folder}.7z"
